% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaev-package.R
\docType{package}
\name{yaev-package}
\alias{yaev-package}
\title{\packageTitle{yaev}}
\description{
The DESCRIPTION file:
\packageDESCRIPTION{yaev}
\packageIndices{yaev}

The \pkg{yaev} package provides utility functions for Extreme
Value Analysis. It includes the probability functions for the
Generalized Pareto Distribution (GPD) and for the Generalized
Extreme Value (GEV) distribution. These functions are vectorized
w.r.t. the parameters and optionnaly provide the exact derivatives
w.r.t. the parameters: gradient and Hessian.

These functions should behave like the probability functions of
the \pkg{stats} package: when a probability \code{p = 0.0} or
\code{p = 1.0} is given, the quantile functions should return the
lower and the upper end-point, be they finite or not. Also when
evaluated at \code{-Inf} and \code{Inf} the probability functions
should return \code{0.0} and \code{1.0}.
}
\section{\pkg{yaev} functions}{
 The \pkg{yaev} functions ...
}

