% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisGP2PP.R
\name{poisGP2PP}
\alias{poisGP2PP}
\title{Transform Poisson-GP Parameters into Point-Process Parameters}
\usage{
poisGP2PP(lambda, loc = 0.0, scale = 1.0, shape = 0.0,
          w = 1.0, deriv = FALSE)
}
\arguments{
\item{lambda}{A numeric vector containing the Poisson rate(s).}

\item{loc}{A numeric vector containing the Generalised Pareto
location, i.e. the threshold in the POT framework.}

\item{scale, shape}{Numeric vectors containing the Generalised Pareto
scale and shape parameters.}

\item{w}{The block duration. Its physical dimension is time and
the product \eqn{\lambda \times w}{\lambda * w} is dimensionless.}

\item{deriv}{Logical. If \code{TRUE} the derivative (Jacobian) of
the transformation is computed and returned as an attribute named
\code{"gradient"} of the attribute.}
}
\value{
A matrix with three columns representing the Point-Process
parameters \code{loc}, \code{scale} and \code{shape}.
}
\description{
Transform Poisson-GP parameters into Point-Process (PP) parameters.
}
\details{
In the POT framework the three parameters are the rate
\code{lambda} of the Poisson process in time: \eqn{\lambda}, and
the two GP parameters: \code{scale} \eqn{\sigma} and \code{shape}
\eqn{\xi}. The vector \code{loc} contains the fixed threshold and
\code{w} the fixed block duration. These parameters are converted
into the vector of three parameters of the GEV distribution for
the maximum of the marks \eqn{Y_i}{Yi} on a time interval with
duration \code{w}, the number \eqn{N} of these marks being a
r.v. with Poisson distribution. More precisely, the GEV
distibution applies when \eqn{N > 0}.
}
\note{
This function is essentially a re-implementation in C of the
function \code{\link[Renext]{Ren2gev}} of \bold{Renext}. As a
major improvement, this function is "vectorised" w.r.t. the
parameters so it can transform efficiently a large number of
Poisson-GP parameter vectors as it can be required e.g. in a MCMC
Bayesian inference. Note also that this function copes with values
near zero for the shape parameter: it suitably computes then both
the function value and its derivatives.
}
\references{
Yves Deville (2020). \emph{Renext Computing Details}. Technical
Report. (Chap. 3 and Appendix A)
}
\seealso{
\code{\link{PP2poisGP}} for the reciprocal
transformation.
}
\author{
Yves Deville
}
